Initialize members of GateTrajectoryNavigator class.
Members were not initialized thus leading to undefined behavior especially for nVerboseLevel.
This class is only used by GateAnalysis and GateToRoot but only GateAnalysis uses the SetVerboseLevel method to explicitly set a value for nVerboseLevel.
On the corresponding class instance used by GateToRoot, nVerboseLevel still had his initial value which is :
-      0 on MacOS 10.5 with libstdc++,
- 131072 (2^16 * 2) on MacOS 10.9 with libc++.

So, the verbose level is unusually and unintentionally high on Mac 10.9 leading to the display of all debugging messages slowing the whole simulation.