use std::abs instead of abs(int) or fabs(float)
clang is complaining when abs and fabs are used with the wrong type,
e.g. abs() when argument is a float, and fabs when argument is an int.
It recommends to always use std::abs, and that seems very reasonable.
(Note that std::abs(arg) is not the same as abs(arg) without the std::
prefix.)