fail on MT geant4, allow user override
Gate does not work correctly with multithreaded Geant4, so if the Geant4
installation appears to be multithreaded then we throw a fatal error.

If the user is a developer who wants to work on making Gate MT-capable,
then this developer can set the GATE_ALLOW_MT_GEANT4 flag, which will
cause cmake to only issue a nanny warning instead of throwing a fatal
error.