add new spot/vertex generation method
Instead of creating a GateSourcePencilBeam objects for each spot
and keeping all of them in a giant vector, create only ONE pencil beam,
and update its parameters when switching between spots.

After reading the source properties and the field/layer/spot specifications,
the total number of primary protons to be generated for the total run is
obtained from the application manager, and then the number of protons in each
of the spots is obtained randomly with the configured spot weights.  This
should yield a distribution that is statistically equivalent to a distribution
obtained with the old method, but it assumes that the 'NumberOfPrimaries' as
set in the application manager refers *only* to a single source.

The code in this commit is NOT yet correct: the GateSourcePencilBeam must
reinitialize itself after setting different parameters. Will take care of that
in the next commit.

The new vertex code is a LOT more user friendly with reading the PBS plan file:
* It allows arbitrary numbers of empty and comment lines.
* It checks the type and the number of items on each given (non-comment) line.
* It tries to print a specific and helpful error message when something is wrong.

Some smaller updates:
* The G4ANALYSIS_USE_ROOT check was superfluous, has been removed.
* Implementation files (.cc) should not need include guards.
* The "number of protons" should be not be an `int` but a `double`, because (1)
  it is really just a weight or an expectation value, and (2) the number returned
  by ConvertMuToNbProtons(w) can be very big, in my example I saw several spots
  with values larger than 5e8, whereas the maximum value that can be stored in an
  `int` is about 2e9, so it is quite possible that for large spot weights an
  `int` is actually insufficient to hold that number computed with
  ConvertMuToNbProtons(w).