make GateSourcePencilBeam re-initializable
Whenever energy or optical parameters are changed, a Pencil Beam source
should be re-initialized. The CLHEP Gaussian random number generators do
not allow to re-set their mean and sigma values, so they need to be
deleted and redefined.  However, the "engine" should not be deleted, so
the correct constructor should be used (the one with the reference
instead of a pointer). The position and rotation input is not needed for
any of the initialization stuff, so they can be changed without
triggering a source re-init.

I actually wonder why this code is not using the static fire/shoot
methods, because those would in fact allow to run with arbitrary mean
and sigma, removing the need for any (re-)initialization.

Smaller stuff:
* The mParticleType data member used to be a character array, now it is
  a with a G4String.
* Trying to be consistent with GateMessage/GateError, avoid G4cout.
* Implementation files do not need include guards.