energy sigma can be absolute (MeV) or relative (%)
If you include a keyword ('percent' or 'MeV') before the polynomial
order for the energy spread in the source properties files, then this
determines whether the value computed from the energy spread polynomial
will be interpreted as an absolute value (in MeV) or as a percentage
(sigma = value*energy/100.).
You can also set a setSigmaEnergyInMeV flag to the TPSPencilBeam actor.
If you specify this choice (absolute or relative) in both places (the
mac file and the source properties file), then the setting in the source
properties file wins. If you specify it in neither place, then "percent"
(relative) is assumed. This is compatible with the wiki docs, but
*incompatible* with the code prior to this commit!