correct sigma energy, remove crashing debug msg
The choice between relative and absolute energy spread specification
should be used not only in the 'old' but also in the 'new' vertex
generation method. The default should be 'relative' (like in the docs)
but without this fix the value from sourceproperties would be used as
absolute (like it was in release 7.2).

I removed a debugging message which tried to use the value of an entry
in the NbProtonsToGenerate vector, but that vector has length 0 in case
the user opts for random spot sampling, which would result in an
embarrassing crash. We don't want that.