use track length in voxel in xsection calc
This is work done by Heide Meissner, a TU Wien student working for/with Hermann Fuchs.
Hermann gave me this code to merge into github.

The main change is to store the total track length inside every voxel and use
that to replace the factor "voxel volume / beam cross section" in the formula
that computes the interesting observable. According to the comments this
enables running the actor for beams under any angle. I suppose that secondary
particles were already running under "any angle" so also for beams parallel to
an axis this improved calculation could be worthwhile.