install only bin/Gate, no (external) headers/libs
Since ages the 'make install' command has installed not only the Gate binary,
but also some random stuff from the itk-mhd external source directory. With this
commit that should be fixed: 'make install' will *only* install the Gate binary
in the $CMAKE_INSTALL_PREFIX/bin directory, nothing else.

To achieve this, I commented something out in one CMakeLists.txt file and set a
flag in another. This way the offending code that was responsible for these
spurious installs is still there, but it is deactivated.  If someone remembers
or stumbles upon a use case in which these lines were actually useful, then we
can re-enable them with the proper conditionals. If this does not happen within
a year or so, then we can just delete this dead code.

This commit should fix issues #75 and #123.