rescale the squared image only if it is enabled
This fixes the segfault that was reported here:
http://lists.opengatecollaboration.org/pipermail/gate-users/2019-January/010639.html

The segfault happened because the dose actor was configured with
"normalize to max", and no "squared image" or uncertainty was enabled.
The code that performs the rescaling / normalizing was trying to rescale
the squared image without checking that it was actually enabled. This
causes a textbook segfault.

I could have made a shorter fix by inserting the
"if (mIsSquaredImageEnabled)"
check twice within the existing 'while' loop, but I prefer to keep long
loops free of conditional statements, which is what the current solution
with two separate while loops achieves.