avoid 'stack smashing' crashes
Using fixed length static C-style character arrays to read the input
from an `istream` is a really bad idea, because `istream` assumes that
the receivers of the input are taking care of their own memory
management. In the case of this commit, if the user configures a
filename path for "/gate/physics/print" that is longer than 30
characters, then the `istream` tries to write beyond the bounds of the
allocated character array. GCC protects against this by placing "canary
strings" directly after a static string, and when it detects changes to
those canary strings then it crashes the program with a 'stack smashing
detected' error.

This is not theoretical, because I have run several times into such
errors, which was quite frustrating.
(I was running lots of simulations with many configurable parameters,
and I used the output directory name to encode the values of those
parameters, which caused those directory names to be much longer than 30
characters.)