GateActorManager: Deactivate previous SD in SetMultiFunctionalDetector
In case a SD is already attached to the logical volume, the SD should be
replaced by a MSD in SetMultiFunctionalDetector. The previously missing
deactivation of the SD could result in a memory leak (see #257).

Since Geant4 doesn't seem to offer a way to remove a SD from the
G4SDManager, we have to deactivate the SD.

Closes #257