efficiency by Z, well defined voxel volume
Two years ago the efficiency was introduced, but it was the same for all
particles. Now you can provide different efficiency tables depending on
the Z of the particles (including Z=-1 for electrons).

The voxel volume used to be undefined for some combinations of
enabled-ness of output images. Now it should always be well defined, for
example if only "dose to water" is requested but not the normal dose.