Fix crash when digitizer noise module used with GateToTree/Singles.
Code relied on "mother" feature of GateHit and GatePulse in order to detect in which systemID pulse or hit are created. But in case of some digitizer module (like GateNoise, pulses are created from scratch, and does not belong to a system). In this case, I choose default system_id = 0.

Maybe, we have to check GateNoise, in order to see if it is correct to create pulses without mother system.