remove c++11 limitation, silence torch warnings
The c++11 requirement was inserted several years ago when some users
were still using OSes with compilers with c++03 as the default standard.
We are now in 2020, the c++11 standard is nine years old and if anyone
still sticks with the 2003 standard then that is not our problem.

Meanwhile, the cmake line that fixes the standard to c++11 is harmful
to code that actually supports newer standards. The c++ standard to
use should be decided by the compiler, not by us.

This issue was triggered by C++ standard warnings in the compilation of
the Gate_NN_ARF_Actor, which relies on Torch. Torch apparently uses C++
idioms from the C++14 and/or C++17 library which resulted in many
warnings. A large fraction of those warnings are now silenced by
removing the `CMAKE_CXX_STANDARD 11` setting.

However, the libtorch code generates are many more compiler warnings,
for example about unused parameters. A good fraction of the warnings are
now silenced by using GCC pragmas in the Gate_NN_ARF_Actor code.  I did
not find pragmas that could silence the many warnings about the
AT_ASSERT and similar macros. That might hopefully be addressed in a
following commit.