avoid segfault in gps energy histogram config
For the GPS config commands it is very important to give them in the
correct order, and this is not always obvious. In case the user tries to
set the histogram interpolation mode before giving (all!) the histogram
points, a segfault would happen. This particular mistake is now caught
and results in an informative error message.

Moreover:
* Almost all commands for the GPS messenger have an old and a new
  version. The old versions are now DEPRECATED, so that in the next
  release we can delete them.
* For the option which confines the source to a volume, the old version had
  smart fixes to the user input, this is now copied to the new version.
* The "ene/gradient" and "ene/emspec" options hardcoded the command name
  to "/gps/ene/gradient" and "/gps/ene/emspec", respectively, instead of
  'directoryname + "ene/gradient"' and 'directoryname + "ene/emspec"', I
  am assuming that that was a typo and fixed it. Needs testing &
  reviewing.