Compile Gate for Linux and Mac after creating the cache
Before, if the cache did not exist, we created it without compile Gate on Linux and MacOs strategy
Now, we compile Gate, because these strategies cannot be run again. To be able to run again, they have to failed but if they failed, they did not store the cache.
For the other strategies, nothing change, we wait the cache to compile Gate