GateActorManager: Replace previous SD in SetMultiFunctionalDetector
Previously, the inactive SD was attached to the
GateMultiSensitiveDetector, leading to an empty hit list. Since the
G4SDStructure member function `RemoveSD` is not exposed and we are
therefore unable to properly remove the original SD from the
G4SDManager, an inactive clone SD is added to the SD manager, which will
implicitly remove the original one. The original and active SD can then
be attached to the multi sensitive detector.

This is a rather dirty workaround and should be removed once `RemoveSD`
or similar is exposed in Geant4.

Note that a Geant4 warning message is raised in case
GEANT4_BUILD_VERBOSE_CODE is set.

Refs #350, #335