Commit to fix #352.
The flag `m_init_done_run_id` was defined as static and did not work for multi system. In this case, one GateDeadTime is created for each system, they can not share the same `m_init_done_run_id`. If they share, `FindLevelsParams` is called only once (for the system which occurs in first pulse).