add ECAT to source/externals
Now the users who want to use ECAT do not need to compile ECAT
separately anymore. When the user enables ECAT (by setting
GATE_USE_ECAT7 to ON), the code under source/externals/ecat will be
built into a libecat.a library and the ECAT parts of Gate are also
compiled and linked into Gate.

We still do not officially maintain the ECAT code. Including this
code in the external section is only done for the convenience of our
users.

TODO: add the original ecat.tar.gz tar ball