add LMF to source/externals
Like ECAT, the LMF library was long distributed by the OpenGATE
collaboration as an unmaintained source tarball, to be built separately
by the interested user. Now we add it to the externals and the build is
done via cmake.

The native LMF built uses a "configure" script which compiles and runs a
short program which generates a small header file with typedefs for
integers with a fixed number of bits. In particular the 32 and 64 bit
integers seem to be a concern due to differences between platforms.
I concocted a little header file myself which uses the <cstdint> header
file to define these typedefs using int32_t, uint32, etc. If some user
gets inspired to become LMF maintainer, using those stdint types instead
of the typedefs would be a nice first cleanup step.

This is just a first commit. I am not an LMF user. An actual LMF user
should try to use the LMF functionality in Gate with this new build
method to check that it indeed works correctly.