Initialize GateSourcePencilBeam::mparticle_time with 0.0
0.0 is the first value that came into my mind, I have no idea
which value should be chosen here.
Initializing the variable fixes a "Conditional jump of move
depends on uninitialized value(s)" message by valgrind, the
fourth finding mentioned in #480.
Some further messages are fixed as well, but there are more
changes than that. Maybe the value of mparticle_time really has an
effect on the control flow?